requirements.txt
annotated-types>=0.5.0
anyio==3.7.1
Authlib==1.2.1
bcrypt>=4.0.1
bleach>=6.0.0
cachetools==5.3.2
certifi==2023.7.22
cffi>=1.15.1
click==8.1.7
contourpy>=1.0.6
cryptography==41.0.7
cycler>=0.11.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.1.3
fastapi>=0.103.2
fonttools>=4.38.0
greenlet==3.0.1
h11==0.14.0
httpcore>=0.17.3
httpx>=0.24.1
idna==3.4
importlib-resources>=5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib>=3.5.3
numpy>=1.21.6
packaging==23.2
pandas>=1.3.5
passlib==1.7.4
Pillow>=9.5.0
plotly==5.18.0
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-settings==2.0.3
pyopenssl>=23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv>=0.21.1
python-multipart==0.0.6
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sse-starlette>=0.10.3
starlette>=0.27.0
tenacity==8.2.3
testresources>=2.0.1
typing-extensions>=4.7.1
tzdata==2023.3
ua-parser==0.18.0
user-agents==2.2.0
uvicorn>=0.22.0
webencodings==0.5.1
zipp>=3.15.0
filelock==3.12.2
APScheduler>=3.10.0
pymysql==1.1.0
psycopg2-binary==2.9.9
lxml==5.1.0
python-jose==3.3.0